SELV3-578: Splitted long cron scheduler into different classes
